Made sure giftnames can be updated

+++ frontend/src/pages/hives/UniqueHive.jsx
@@ -3,6 +3,8 @@ import { useGiftStore } from "../../stores/useGiftStore";
 import { useParams } from "react-router-dom";
 import { Link } from "react-router-dom";
 import { Button } from "../../components/elements/Button/Button";
+import update from "../../assets/update.svg"
+import Swal from "sweetalert2";
 import trashcanIcon from "../../assets/trash.svg";
 
 
@@ -14,7 +16,7 @@ const withEndpoint = (endpoint) => `${API_URL}/gift-routes/${endpoint}`;
 export const UniqueHive = () => {
     const { id } = useParams();
     const [hive, setHive] = useState(null);
-    const { getHives, deleteGift } = useGiftStore();
+    const { getHives, deleteGift, updateGift } = useGiftStore();
 
     useEffect(() => {
         const handleHiveFetch = async () => {
@@ -39,31 +41,75 @@ export const UniqueHive = () => {
                 console.error("There was an error =>", error);
             }
         };
-
-        // Call the fetchMovie function when the component mounts or when "id" changes
-        // The "id" parameter is included as a dependency in the array to trigger a re-fetch, useEffect will be called when "id" changes.
         handleHiveFetch();
     }, [id, hive]);
 
-    const handleDelete = async (hiveId) => {
+    const handleUpdateGift = async (giftId) => {
+        // const newHiveName = prompt("Enter new hive name");
+        const { value: updatedGift } = await Swal.fire({
+            title: "Update gift",
+            input: "text",
+            inputLabel: "What would you like to change your gifts name to be(e)? 🐝",
+            inputPlaceholder: "e.g. Mom",
+            showCancelButton: true,
+            inputValidator: (value) => {
+                if (!value) {
+                    return "You need to write something!";
+                }
+            },
+        });
 
-        try {
-            await deleteGift(hiveId);
-            getHives();
-        } catch (error) {
-            console.error("There was an error =>", error);
+        if (updatedGift) {
+            try {
+                await updateGift({ id: giftId, gift: updatedGift, tags: [], bought: false });
+                getHives();
+            } catch (error) {
+                console.error("There was an error =>", error);
+                Swal.fire({
+                    icon: "error",
+                    title: "Oops...",
+                    text: "Something went wrong!",
+                });
+            }
         }
+    };
+
+    const handleDelete = async (hiveId) => {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "Yes, delete it!",
+            cancelButtonText: "No, keep it",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                try {
+                    deleteGift(hiveId);
+                    getHives();
+                } catch (error) {
+                    console.error("There was an error =>", error);
+                }
+            } else if (result.dismiss === Swal.DismissReason.cancel) {
+                Swal.fire({
+                    title: "Cancelled",
+                    text: "Your gift is safe 🐝",
+                    icon: "error"
+                })
+            }
+        });
     };
 
     return (
         <section>
             {hive && hive.gifts ? (
                 <div>
-                    <h2>{hive.name}</h2>
+                    <h1>{hive.name}</h1>
                     {hive.gifts.map((gift) => {
                         return (
                             <ul className="list-item-pair" key={gift._id}>
                                 <li>{gift.gift}</li>
+                                <img className="icon" src={update} alt="Icon for updating the hives name" onClick={() => handleUpdateGift(gift._id)} />
                                 <img className="icon" src={trashcanIcon} alt="Trashcan for deleting a hive" onClick={() => handleDelete(gift._id)} />
                             </ul>
                         );
